SearchTarget周りの実装とテストを分割
