+++ src/lib/search-target/index.ts
@@ -1,0 +1,3 @@
+export * from './key';
+export * from './label';
+export * from './search-target';
