+++ src/lib/search-target/search-target.ts
@@ -1,4 +1,4 @@
-import { Tag } from './ship';
+import { Tag } from '../ship';
 
 export enum SearchType {
   TAG,
@@ -15,14 +15,6 @@ interface SearchTargetTagCategory {
 }
 
 export type SearchTarget = SearchTargetName | SearchTargetTagCategory;
-
-export const getKeyOfSearchTarget = (target: SearchTarget) => {
-  return target.type === SearchType.NAME ? 'name' : target.category;
-};
-
-export const getLabelOfSearchTarget = (target: SearchTarget) => {
-  return target.type === SearchType.NAME ? '名前' : target.category;
-};
 
 export const generateSearchTargets = (tags: Tag[]): SearchTarget[] => {
   const categories = new Set<string>(tags.map(({ category }) => category));
